pyproject.toml
[project]
name = "llm-auto-api-integration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.77.0",
    "python-dotenv>=1.0.0",
    "PyQt6>=6.6.1",
]
